fix: fixing empty cart
